parse_helpers for conn
