+++ bro_conn_parse.lua
@@ -1,40 +1,4 @@
-function string:toboolean()
-  --- constants
-  local TRUE = {
-      ['1'] = true,
-      ['t'] = true,
-      ['T'] = true,
-      ['true'] = true,
-      ['TRUE'] = true,
-      ['True'] = true,
-  }
-  local FALSE = {
-      ['0'] = false,
-      ['f'] = false,
-      ['F'] = false,
-      ['false'] = false,
-      ['FALSE'] = false,
-      ['False'] = false,
-  }
-  
-  if TRUE[self] == true then
-      return true
-  elseif FALSE[self] == false then
-      return false
-  else
-      return nil
-  end
-end
-
-function variable_to_boolean(var)
-  if type(var) == 'string' then
-    return var:toboolean()
-  elseif type(var) == 'boolean' then
-    return var
-  else
-    return nil
-  end
-end
+require('parse_helpers')
 
 function bro_conn_parse_direction(tag, timestamp, record)
   local local_orig = variable_to_boolean(record["local_orig"])
