+++ fluent-bit-test.start
@@ -4,4 +4,6 @@
 # See https://github.com/fluent/fluent-bit/issues/326
 export TZ=UTC
 
+export LUA_PATH="$HOME/fluent-bit-test/?.lua;"
+
 fluent-bit -c ~/fluent-bit-test/fluent-bit-test.conf
